% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_sites_years_occ.R
\name{define_sites_years_occ}
\alias{define_sites_years_occ}
\title{Create n sites sampled over t years}
\usage{
define_sites_years_occ(pool, n_years, n_sites, rich_mean = 30, rich_sd = 5)
}
\arguments{
\item{pool}{list, a list generated by sp_pool() containing species and ditrib}

\item{n_years}{Numeric, number of years}

\item{n_sites}{Numeric, number of sites (imagine a lattice)}

\item{rich_mean}{Numeric, mean species richness expected per site.}

\item{rich_sd}{Numeric, standard deviation of the species richness expected per site}
}
\value{
A data.frame.
}
\description{
This function creates for a given species pool, a data.frame of species observed per site and year.
}
\details{
Species are
assigned to each site randomly according to its expected distribution, ensuring
that common species appear on more sites than rare species. Richness in not controled
and scales with number of species. Note that we
assume no immigration / emigration.
}
\examples{
pool <- sp_pool(50)
define_sites_years_occ(pool = pool, n_years = 3, n_sites = 10)
}
