% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_ifelse.R
\name{fast_ifelse}
\alias{fast_ifelse}
\title{fast_ifelse}
\usage{
fast_ifelse(test, yes, no)
}
\arguments{
\item{test}{to test}

\item{yes}{output if yes}

\item{no}{output if no}
}
\description{
fast_ifelse
}
