% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_abundance.R
\name{obs_abundance}
\alias{obs_abundance}
\title{Create observed abundances with detection error per site and year}
\usage{
obs_abundance(true_abundance, sp_responses)
}
\arguments{
\item{true_abundance}{A data.frame, generated by true_abundances()}

\item{sp_responses}{A data.frame, generated by sp_responses()}
}
\value{
A vector or data.frame.
}
\description{
This function simulate sampling with a detection probablity per species.
}
\examples{
obs_abundance(true_abundance, sp_responses)
}
