% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rspecies.R
\name{rspecies}
\alias{rspecies}
\title{Rspecies}
\usage{
rspecies(
  n,
  spp,
  b = rep(0, spp),
  x = rep(1, n),
  type = c("glm", "glm.nb", "glmm"),
  family = gaussian,
  sigma = 1,
  cor = diag(1, n, n),
  theta = 1,
  attrib = TRUE,
  seed = NULL
)
}
\arguments{
\item{n}{See original funtcion}

\item{spp}{See original funtcion}

\item{b}{See original funtcion}

\item{x}{See original funtcion}

\item{type}{See original funtcion}

\item{family}{See original funtcion}

\item{sigma}{See original funtcion}

\item{cor}{See original funtcion}

\item{theta}{See original funtcion}

\item{attrib}{See original funtcion}

\item{seed}{See original funtcion}
}
\value{
A matrix.
}
\description{
This function is used from https://rdrr.io/rforge/rspecies/man/rspecies.html. See help there.
}
