% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPoly-package.R
\docType{package}
\name{SimPoly-package}
\alias{SimPoly}
\alias{SimPoly-package}
\title{SimPoly: Like SimCity but for pollinator communities}
\description{
Pollinator community simulator to implment power anlysis in EU-PoMS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ibartomeus/SimPoly}
  \item Report bugs at \url{https://github.com/ibartomeus/SimPoly/issues}
}

}
\author{
\strong{Maintainer}: Ignasi Bartomeus \email{nacho.bartomeus@gmail.com} (\href{https://orcid.org/0000-0001-7893-4389 }{ORCID})

Other contributors:
\itemize{
  \item Nick Isaac [contributor]
  \item Oliver Schweiger [contributor]
}

}
\keyword{internal}
