% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_pool.R
\name{sp_pool}
\alias{sp_pool}
\title{Set species pool}
\usage{
sp_pool(pool = 100, rarest = 0.05, commonest = 0.7)
}
\arguments{
\item{pool}{Numeric, number of species}

\item{rarest}{Numeric, true occupancy of the rarest species from 0-1}

\item{commonest}{Numeric, true occupancy of the commonest species from 0-1}
}
\value{
A list with a vector of species, and a vector of its distribution.
}
\description{
This function sets the species pool for a given region
}
\examples{
sp_pool(pool = 50, rarest = 0.05, commonest = 0.7)
}
