% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_poms.R
\name{summary_poms}
\alias{summary_poms}
\title{Create a summary by year species and site}
\usage{
summary_poms(eupoms, var_name = "obs")
}
\arguments{
\item{eupoms}{Data.frame, generated by true_abundances() and obs_abundance()}

\item{var_name}{String, the name of the variable storing the observed abundances, which should a be numeric variable.}
}
\value{
A data.frame.
}
\description{
This aggregates all rounds per year species and site
}
\examples{
summary_poms(eupoms)
}
