% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_sites_years.R
\name{define_sites_years}
\alias{define_sites_years}
\title{Create n sites sampled over t years}
\usage{
define_sites_years(nsp, n_years, n_sites, ctrl = -3, heterogeneity = 1)
}
\arguments{
\item{nsp}{Numeric, number of species in the region.}

\item{n_years}{Numeric, number of years}

\item{n_sites}{Numeric, number of sites (imagine a lattice)}

\item{ctrl}{Numeric, tunes the richness per site. has to be negative, the lower, the less richness.}

\item{heterogeneity}{A number of environmental heterogeneity expected across sites.
It can be read as an SD across MS with LX close to 0 and Spain ~1. If MS values are in another scale,
we can remove the minimum and divide by max to get 0-1 scale}
}
\value{
A data.frame.
}
\description{
This function creates for a given species pool, a data.frame of species observed per site and year.
}
\details{
We use the function rspecies to create a matrix of sites per species, where each species occurrs
in different sites according to their distribution (asumer normal occupancies) and sites show more or less similar
community composition depending on an envoronmental gradient. Note that we
assume no immigration / emigration.
}
\examples{
define_sites_years(nsp = 100, n_years = 3, n_sites = 10, heterogeneity = 1)
}
