% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_abundance.R
\name{obs_abundance}
\alias{obs_abundance}
\title{Create observed abundances with detection error per site and year}
\usage{
obs_abundance(
  true_abundance = NULL,
  sp_responses,
  fraction_observed = 0.5,
  noise_frac = 0.1,
  pantrap = TRUE,
  n_pan = 5
)
}
\arguments{
\item{true_abundance}{A data.frame, generated by true_abundances()}

\item{sp_responses}{A data.frame, generated by sp_responses()}

\item{fraction_observed}{numeric, the fraction of individuals present this day that can be observed. Default 0.5}

\item{noise_frac}{numeric, noise around fraction observed. Default 0.1}

\item{pantrap}{Logical, should pantraps be included? Default TRUE}

\item{n_pan}{numeric, number of pantraps deployed. Default 5}
}
\value{
A vector or data.frame.
}
\description{
This function simulate sampling with a detection probability per species.
}
\details{
For each species, site, year and round combination, individuals are sampled without replacement randomly
proportional to its abundance, but weigthed by their probability of detection. the faction observed has built in
noise accounting for "good nad "bad" years. Pantrap observations are
modeled using a binomial distribution where each individual present has a probaility to falling/not falling
in each pantrap. We report for each species, site, year and round, how many pantraps catch a give species.
}
\note{
This function is slow.
}
\examples{
site_years <- define_sites_years(nsp = 100, n_years = 3, n_sites = 10)
pars <- sp_responses(site_years = site_years)
abun <- true_abundance(n_round = 8,
                                 site_years = site_years,
                                 sp_responses = pars)
obs_abundance(true_abundance = abun, sp_responses = pars)

}
