% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_abundance.R
\name{obs_abundance}
\alias{obs_abundance}
\title{Create observed abundances with detection error per site and year}
\usage{
obs_abundance(
  true_abundance = NULL,
  sp_responses,
  fraction_observed = 0.5,
  pantrap = TRUE,
  n_pan = 5
)
}
\arguments{
\item{true_abundance}{A data.frame, generated by true_abundances()}

\item{sp_responses}{A data.frame, generated by sp_responses()}

\item{fraction_observed}{numeric, thefraction of individuals present this day that can be observed. Default 0.5}

\item{pantrap}{Logical, should pantraps be included? Default TRUE}

\item{n_pan}{numeric, number of pantraps deployed. Default 5}
}
\value{
A vector or data.frame.
}
\description{
This function simulate sampling with a detection probablity per species.
}
\examples{
pool <- sp_pool(50)
site_years <- define_sites_years(pool = pool, n_years = 3, n_sites = 10)
pars <- sp_responses(site_years = site_years)
abun <- true_abundance(rounds = 8,
                                 site_years = site_years,
                                 sp_responses = pars)
obs_abundance(true_abundance = abun, sp_responses = pars)

}
