% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_responses.R
\name{sp_responses}
\alias{sp_responses}
\title{Create species responses}
\usage{
sp_responses(
  site_years,
  pheno_peak_mean = 120,
  pheno_peak_sd = 50,
  pheno_range_mean = 25,
  pheno_range_sd = 5,
  trend_max = 1.1,
  trend_min = 0.9
)
}
\arguments{
\item{site_years}{A data.frame, generated by define_sites_years()}

\item{pheno_peak_mean}{Numeric, mean of the species phenology peak (in days from 1-365)}

\item{pheno_peak_sd}{Numeric, sd of the species phenology peak}

\item{pheno_range_mean}{Numeric, mean of the species phenology ranges}

\item{pheno_range_sd}{Numeric, sd of the species phenology ranges}

\item{trend_max}{Numeric, minimum slope of linear decline per year. Take values between 0 and infinite, with values < 1 implying declines.}

\item{trend_min}{Numeric, máximum slope of linear decline per year. Take values between 0 and infinite, with values < 1 implying declines.}
}
\value{
A data.frame.
}
\description{
This function defines for each species its phenology, typical abundance, detectablity  and temporal trend.
}
\examples{
pool <- sp_pool(50)
site_years <- define_sites_years(pool = pool, n_years = 3, n_sites = 10)
sp_responses(site_years = site_years)
}
