% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_poms.R
\name{summary_poms}
\alias{summary_poms}
\title{Create a summary by year species and site}
\usage{
summary_poms(eupoms, var_name = "obs")
}
\arguments{
\item{eupoms}{Data.frame, generated by true_abundances() and obs_abundance()}

\item{var_name}{String, the name of the variable storing the observed abundances, which should a be numeric variable.}
}
\value{
A data.frame.
}
\description{
This aggregates all rounds per year species and site
}
\examples{
pool <- sp_pool(50)
site_years <- define_sites_years(pool = pool, n_years = 3, n_sites = 10)
pars <- sp_responses(site_years = site_years)
true_abundance <- true_abundance(n_round = 8,
                         site_years = site_years,
                         sp_responses = pars)
eupoms <- obs_abundance(true_abundance = true_abundance, sp_responses = pars)
summary_poms(eupoms)
}
