% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_abundance.R
\name{true_abundance}
\alias{true_abundance}
\title{Create true abundances per site and year}
\usage{
true_abundance(
  rounds = 8,
  startmonth = 2,
  endmonth = 10,
  site_years = NA,
  sp_responses = NA
)
}
\arguments{
\item{rounds}{Numeric. Number of sampling rounds. Default to 8.}

\item{startmonth}{Numeric. Number of month starting to sample. Default to 2.}

\item{endmonth}{Numeric. Number of month stopping to sample. Default to 10.}

\item{site_years}{A data.frame, generated by define_sites_years()}

\item{sp_responses}{A data.frame, generated by sp_responses()}
}
\value{
A data.frame.
}
\description{
This function calculate true abundance in each sampling round.
}
\examples{
pool <- sp_pool(50)
site_years <- define_sites_years(pool = pool, n_years = 3, n_sites = 10)
pars <- sp_responses(site_years = site_years)
true_abundance(rounds = 8,
               site_years = site_years,
               sp_responses = pars)
}
